Added --older-than flag in mlflow gc command, and added tests for that cli command

+++ mlflow/cli.py
@@ -1,10 +1,13 @@
 import json
 import os
+import time
+import re
 import sys
 import logging
 
 import click
 from click import UsageError
+from datetime import datetime, timedelta
 
 import mlflow.db
 import mlflow.experiments
@@ -484,6 +487,14 @@ def server(
 
 @cli.command(short_help="Permanently delete runs in the `deleted` lifecycle stage.")
 @click.option(
+    "--older-than",
+    default="0",
+    help="Optional Remove run(s) older than the specified time limit. "
+    "Specify a string in #d#h#m#s format. "
+    "For example: --older-than 1d2h3m4s"
+    "Defaults to 0 (all runs).",
+)
+@click.option(
     "--backend-store-uri",
     metavar="PATH",
     default=DEFAULT_LOCAL_FILE_AND_ARTIFACT_PATH,
@@ -500,7 +511,7 @@ def server(
     " are not specified, data is removed for all runs in the `deleted`"
     " lifecycle stage.",
 )
-def gc(backend_store_uri, run_ids):
+def gc(older_than, backend_store_uri, run_ids):
     """
     Permanently delete runs in the `deleted` lifecycle stage from the specified backend store.
     This command deletes all artifacts and metadata associated with the specified runs.
@@ -510,8 +521,35 @@ def gc(backend_store_uri, run_ids):
         raise MlflowException(
             "This cli can only be used with a backend that allows hard-deleting runs"
         )
+
+    current_time = int(time.time() * 1000)
+
+    if older_than != "0":
+        regex = re.compile(
+            r"^((?P<days>[\.\d]+?)d)?((?P<hours>[\.\d]+?)h)?((?P<minutes>[\.\d]+?)m)?((?P<seconds>[\.\d]+?)s)?$"
+        )
+        parts = regex.match(older_than)
+        if parts is None:
+            raise MlflowException(
+                "Could not parse any time information from '{}'.  Examples of valid strings: '8h', '2d8h5m20s', '2m4s'".format(
+                    older_than
+                )
+            )
+        time_params = {name: float(param) for name, param in parts.groupdict().items() if param}
+        time_delta = int(timedelta(**time_params).total_seconds() * 1000)
+
     if not run_ids:
-        run_ids = backend_store._get_deleted_runs()
+        if older_than != "0":
+            # filter run_ids based on older_than parameter
+            _run_ids = backend_store._get_deleted_runs()
+            run_ids = []
+            for run_id in _run_ids:
+                run = backend_store.get_run(run_id)
+                if current_time - run.info.delete_time > time_delta:
+                    run_ids.append(run_id)
+        else:
+            run_ids = backend_store._get_deleted_runs()
+
     else:
         run_ids = run_ids.split(",")
 
@@ -521,6 +559,11 @@ def gc(backend_store_uri, run_ids):
             raise MlflowException(
                 "Run % is not in `deleted` lifecycle stage. Only runs in "
                 "`deleted` lifecycle stage can be deleted." % run_id
+            )
+        # raise MlflowException if run_id is newer than older_than parameter
+        if older_than != "0" and current_time - run.info.delete_time <= time_delta:
+            raise MlflowException(
+                f"Run {run_id} is not older than the required age. Only runs older than {older_than} can be deleted."
             )
         artifact_repo = get_artifact_repository(run.info.artifact_uri)
         artifact_repo.delete_artifacts()
